% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{fast_mgcv_plot_data}
\alias{fast_mgcv_plot_data}
\title{Create data for plotting all the smooth effects of an mgcv object}
\usage{
fast_mgcv_plot_data(
  fit,
  n = 1000,
  coord_names = c("lon", "lat"),
  lon_range = NULL,
  lat_range = NULL
)
}
\description{
Create data for plotting all the smooth effects of an mgcv object
}
