% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{fast_mgcv_pred}
\alias{fast_mgcv_pred}
\title{Perform fast prediction for gam objects with smooths and an intercept}
\usage{
fast_mgcv_pred(fit, data, detailed = FALSE)
}
\description{
Perform fast prediction for gam objects with smooths and an intercept
}
