% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{iqd}
\alias{iqd}
\title{Estimate the (possibly weighted) IQD between x and y, using numerical integration}
\usage{
iqd(x, y, w = function(x) 1, rm_zero = FALSE, zero_threshold = 0)
}
\description{
Estimate the (possibly weighted) IQD between x and y, using numerical integration
}
