% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_station_data.R
\name{load_station_data}
\alias{load_station_data}
\title{Load weather station data, ERA data and covariate data from a set of locations}
\usage{
load_station_data(
  meta,
  station_dir,
  verbose = FALSE,
  rm_bad_flags = FALSE,
  bad_flag_start = "H",
  rm_na = TRUE,
  era_stats = FALSE,
  rm_leading_na = TRUE
)
}
\description{
\subsection{Precipitation flag definition
(from https://www.ncei.noaa.gov/data/global-summary-of-the-day/doc/readme.txt)}{

A = 1 report of 6-hour precipitation amount.
B = Summation of 2 reports of 6-hour precipitation amount.
C = Summation of 3 reports of 6-hour precipitation amount.
D = Summation of 4 reports of 6-hour precipitation amount.
E = 1 report of 12-hour precipitation amount.
F = Summation of 2 reports of 12-hour precipitation amount.
G = 1 report of 24-hour precipitation amount.
H = Station reported '0' as the amount for the day (eg, from 6-hour reports),
but also reported at least one occurrence of precipitation in hourly observations.
This could indicate a trace occurred, but should be considered as incomplete
data for the day.
I = Station did not report any precipitation data for the day and did not report any
occurrences of precipitation in its hourly observations. It's still possible that
precipitation occurred but was not reported.
}
}
