% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{opposite_skill_score}
\alias{opposite_skill_score}
\title{Function for computing what the skill score of (s0, s1) would have been, when we
know what the skill score, s, of (s1, s0) is}
\usage{
opposite_skill_score(s)
}
\description{
Function for computing what the skill score of (s0, s1) would have been, when we
know what the skill score, s, of (s1, s0) is
}
